get_changelogs_in_date_range out of fab package
